pyproject.toml
[tool.poetry]
name = "labby"
version = "0.1.0"
description = "CLI tool to build Network Labs in an automated way"
authors = ["David Flores <davidflores7_8@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
gns3fy = "^0.7.1"
typer = "^0.3.2"
rich = "^9.1.0"
nornir = "^3.0.0"
scrapli = "^2020.10.10"
nornir-scrapli = "^2020.11.1"
tomlkit = "^0.7.0"
pydantic = "^1.7.2"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
black = "^20.8b1"
flake8 = "^3.8.4"

[tool.poetry.scripts]
labby = "labby.main:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
